% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cand_points.R
\name{get_cand_points}
\alias{get_cand_points}
\title{Create dataframe of candidate points.}
\usage{
get_cand_points(study_frame, n, pt_density = 10)
}
\arguments{
\item{study_frame}{Sf dataframe of the sample area.}

\item{n}{Number of sites in the final monitoring network.}

\item{pt_density}{Density of potential points on the landscape.}
}
\value{
sf dataframe of potential sites
}
\description{
Create dataframe of candidate points.
}
