% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eqprob.R
\name{get_eqprob}
\alias{get_eqprob}
\title{Equal probability monitoring networks}
\usage{
get_eqprob(sf_df, n)
}
\arguments{
\item{sf_df}{sf dataframe defining the sample frame}

\item{n}{number of sites to be selected}
}
\value{
list of point indexes for each algorithm
}
\description{
Equal probability monitoring networks
}
